requirements.txt
configparser~=5.0.1
requests~=2.25.1
gTTS~=2.2.2
playsound~=1.2.2
python-dateutil~=2.8.1
art~=5.1
termcolor~=1.1.0
SQLAlchemy~=1.3.23
mysqlclient~=2.0.3
mysql-connector-python~=8.0.23
fastapi~=0.63.0